renamed message_producer.controller as initializer
